Low V Fit Opt + Fit Graph Formatting + Fitting Parameters as Arguments
All fitting parameters are now arguments in fit_atlung.

Fit graphs are from 1e-6 to 1e2 and have label in the top left.

Calculates without points with fcap less than 0.001 if the largest fcap is less than 0.1. This allows better fits by treating the resistance growth at low voltages as an instantaneous ohmic resistance when total fcap is small enough that this makes a difference in the fit.